# server/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 根据ID获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 根据ID获取用户信息
      tags:
      - 用户管理
    patch:
      consumes:
      - application/json
      description: 更新现有用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 更新用户信息
      tags:
      - 用户管理
  /users/export:
    post:
      consumes:
      - application/json
      description: 导出用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 导出用户
      tags:
      - 用户管理
...
  /users/role/{id}:
    get:
      consumes:
      - application/json
      description: 查询用户（查询某个角色下的所有用户）
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 查询用户（查询某个角色下的所有用户）
      tags:
      - 用户管理
swagger: "2.0"